Splitting tests and creating new autobuild @rev mohamed-m@ (#147)
* Splitting tests and creating new autobuild

* Minor update to circleci config

* Fixing issues

* Reducing coverage threshold

* Adding base tests to workflows

* Renaming test and adding docstring
