+++ python/ml4ir/base/tests/test_ml4ir_sanity.py
@@ -118,13 +118,8 @@ class TestML4IRSanity(unittest.TestCase):
                         log_regression_mrr=1.0,
                         working_dir=pathlib.Path(self.working_dir.path), log_dir=pathlib.Path(self.log_dir))
 
-    def _test_linear_ml4ir_sanity_2(self):
+    def test_linear_ml4ir_sanity_2(self):
         """
-        Note
-        ----
-        Disabling temporarily as the CircleCI builds are hitting OutOfMemory
-        due to a tensorflow-keras memory leakage issue.
-
         Weights of pre-trained models:
             perceptron=np.array([4.50209484, -0.80280452])
             log_reg=np.array([22.73585585, -3.94821153])
@@ -145,13 +140,8 @@ class TestML4IRSanity(unittest.TestCase):
                         log_regression_mrr=1.0,
                         working_dir=pathlib.Path(self.working_dir.path), log_dir=pathlib.Path(self.log_dir))
 
-    def _test_linear_ml4ir_sanity_4(self):
+    def test_linear_ml4ir_sanity_4(self):
         """
-        Note
-        ----
-        Disabling temporarily as the CircleCI builds are hitting OutOfMemory
-        due to a tensorflow-keras memory leakage issue.
-
         Weights of pre-trained models:
             perceptron=np.array([5.10535665, 1.44131417])
             log_reg=np.array([20.0954756, 4.69360163])
@@ -172,13 +162,8 @@ class TestML4IRSanity(unittest.TestCase):
                         log_regression_mrr=0.4290204375570229,
                         working_dir=pathlib.Path(self.working_dir.path), log_dir=pathlib.Path(self.log_dir))
 
-    def _test_linear_ml4ir_sanity_6(self):
+    def test_linear_ml4ir_sanity_6(self):
         """
-        Note
-        ----
-        Disabling temporarily as the CircleCI builds are hitting OutOfMemory
-        due to a tensorflow-keras memory leakage issue.
-
         Weights of pre-trained models:
         perceptron=np.array(
                         [4.59994733, -3.56373965, -6.15935686, 0.87523846, -0.64231058, 2.15971991, 5.79875003,
@@ -203,14 +188,8 @@ class TestML4IRSanity(unittest.TestCase):
                         log_regression_mrr=0.4938108725899423,
                         working_dir=pathlib.Path(self.working_dir.path), log_dir=pathlib.Path(self.log_dir))
 
-    def _test_linear_ml4ir_sanity_8(self):
+    def test_linear_ml4ir_sanity_8(self):
         """
-        Note
-        ----
-        Disabling temporarily as the CircleCI builds are hitting OutOfMemory
-        due to a tensorflow-keras memory leakage issue.
-
-        
         Weights of pre-trained models:
             perceptron=np.array(
                             [2.91798129, 4.24880336, 7.42919018, 2.49609694, -0.84988373, 0.43435823, -0.18953416,
